Viz: Fix path pretty printing bug (path graphs had to be (re-?)augmented with edge labels)
